refactor(kernel): complete integration test migration to external tests/
Move 27 integration tests from src/tests.rs to tests/kernel_integration.rs.
Keep 16 unit tests (Fault, TaskState FSM, proptest, pack validation,
token generation, task supervisor rejects, record_tool_call_denial)
inline where they test single components or need pub(crate) access.
